Refactor modal structure and update import paths for improved organization
This commit reorganizes the modal components by consolidating imports into a centralized structure under the modals directory. It updates the import paths across various components, including EyeWorkerService and GazePlotter, to enhance clarity and maintainability. Additionally, new modal content components have been introduced, streamlining the modal functionality and improving the overall project structure.
