+++ src/lib/components/GazePlotter.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Toaster from '$lib/components/Toaster/Toaster.svelte'
-  import Modal from '$lib/components/Modal/Modal.svelte'
+  import { Modal } from '$lib/modals'
   import Panel from '$lib/components/Panel/Panel.svelte'
   import Workspace from '$lib/components/Workspace/Workspace.svelte'
   import Tooltip from './Tooltip.svelte'
