--- src/lib/components/Modal/ModalContent/index.ts
@@ -1,17 +1,0 @@
-// Modal Content Components
-export { default as ModalContentMaxValue } from './ModalContentMaxValue.svelte'
-export { default as ModalContentColorScale } from './ModalContentColorScale.svelte'
-export { default as ModalContentAoiVisibility } from './ModalContentAoiVisibility.svelte'
-export { default as ModalContentAoiModification } from './ModalContentAoiModification.svelte'
-export { default as ModalContentParticipantModification } from './ModalContentParticipantModification.svelte'
-export { default as ModalContentStimulusModification } from './ModalContentStimulusModification.svelte'
-export { default as ModalContentParticipantsGroups } from './ModalContentParticipantsGroups.svelte'
-export { default as ModalContentParticipantsGroupsChecklist } from './ModalContentParticipantsGroupsChecklist.svelte'
-export { default as ModalContentBarChartAxes } from './ModalContentBarChartAxes.svelte'
-export { default as ModalContentScarfPlotClip } from './ModalContentScarfPlotClip.svelte'
-export { default as ModalContentTobiiParsingInput } from './ModalContentTobiiParsingInput.svelte'
-export { default as ModalContentDownloadWorkplace } from './ModalContentDownloadWorkplace.svelte'
-export { default as ModalContentDownloadBarPlot } from './ModalContentDownloadBarPlot.svelte'
-export { default as ModalContentDownloadScanGraph } from './ModalContentDownloadScanGraph.svelte'
-export { default as ModalContentDownloadScarfPlot } from './ModalContentDownloadScarfPlot.svelte'
-export { default as ModalContentDownloadTransitionMatrix } from './ModalContentDownloadTransitionMatrix.svelte'
