--- src/lib/components/Modal/index.ts
@@ -1,5 +1,0 @@
-// Main Modal Component
-export { default as Modal } from './Modal.svelte'
-
-// Modal Content Components
-export * from './ModalContent'
