+++ src/lib/components/Panel/PanelButton/PanelButtonExportWorkplace.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import MajorControlButton from '$lib/components/General/GeneralButton/GeneralButtonMajor.svelte'
-  import { modalStore } from '$lib/stores/modalStore'
-  import ModalContentDownloadWorkplace from '$lib/components/Modal/ModalContent/ModalContentDownloadWorkplace.svelte'
+  import { modalStore } from '$lib/modals/shared/stores/modalStore'
+  import { ModalContentDownloadWorkplace } from '$lib/modals'
   import { processingFileStateStore } from '$lib/stores/processingFileStateStore'
 
   let isDisabled = $derived($processingFileStateStore === 'processing')
 
   const handleClick = () => {
-    modalStore.open(ModalContentDownloadWorkplace, 'Download workplace')
+    modalStore.open(ModalContentDownloadWorkplace as any, 'Download workplace')
   }
 </script>
