+++ src/lib/components/Panel/PanelButton/PanelButtonScanGraph.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import MajorControlButton from '$lib/components/General/GeneralButton/GeneralButtonMajor.svelte'
-  import { modalStore } from '$lib/stores/modalStore.js'
-  import ModalContentDownloadScanGraph from '$lib/components/Modal/ModalContent/ModalContentDownloadScanGraph.svelte'
+  import { modalStore } from '$lib/modals/shared/stores/modalStore.js'
+  import { ModalContentDownloadScanGraph } from '$lib/modals'
   import { processingFileStateStore } from '$lib/stores/processingFileStateStore'
 
   let isDisabled = $derived($processingFileStateStore === 'processing')
 
   const handleClick = () => {
-    modalStore.open(ModalContentDownloadScanGraph, 'Export to ScanGraph')
+    modalStore.open(ModalContentDownloadScanGraph as any, 'Export to ScanGraph')
   }
 </script>
