+++ src/lib/modals/import/components/ModalContentAoiVisibility.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-  import GeneralSelectBase from '../../General/GeneralSelect/GeneralSelect.svelte'
+  import GeneralSelectBase from '$lib/components/General/GeneralSelect/GeneralSelect.svelte'
   import { getParticipants } from '$lib/stores/dataStore.js'
-  import GeneralInputFile from '../../General/GeneralInput/GeneralInputFile.svelte'
-  import GeneralButtonMajor from '../../General/GeneralButton/GeneralButtonMajor.svelte'
+  import GeneralInputFile from '$lib/components/General/GeneralInput/GeneralInputFile.svelte'
+  import GeneralButtonMajor from '$lib/components/General/GeneralButton/GeneralButtonMajor.svelte'
   import { addErrorToast, addSuccessToast } from '$lib/stores/toastStore.js'
-  import { processAoiVisibility } from '$lib/services/aoiVisibilityServices'
+  import { processAoiVisibility } from '$lib/modals/import/utility/aoiVisibilityServices'
   import { getStimuliOptions } from '$lib/plots/shared/utils/sharedPlotUtils'
   interface Props {
     forceRedraw: () => void
