+++ src/lib/modals/import/components/index.ts
@@ -1,0 +1,3 @@
+// Import Components
+export { default as ModalContentTobiiParsingInput } from './ModalContentTobiiParsingInput.svelte'
+export { default as ModalContentAoiVisibility } from './ModalContentAoiVisibility.svelte'
