+++ src/lib/modals/modification/components/ModalContentBarChartAxes.svelte
