+++ src/lib/modals/modification/components/ModalContentParticipantModification.svelte
@@ -2,8 +2,7 @@
   import GeneralButtonMajor from '$lib/components/General/GeneralButton/GeneralButtonMajor.svelte'
   import GeneralButtonPreset from '$lib/components/General/GeneralButton/GeneralButtonPreset.svelte'
   import GeneralInputText from '$lib/components/General/GeneralInput/GeneralInputText.svelte'
-  import SortableTableHeader from '$lib/components/Modal/Shared/SortableTableHeader.svelte'
-  import SectionHeader from '$lib/components/Modal/Shared/SectionHeader.svelte'
+  import { SortableTableHeader, SectionHeader } from '$lib/modals'
   import {
     getAllParticipants,
     updateMultipleParticipants,
