+++ src/lib/modals/modification/components/index.ts
@@ -1,0 +1,12 @@
+// Visualization Settings Modals
+export { default as ModalContentMaxValue } from './ModalContentMaxValue.svelte'
+export { default as ModalContentColorScale } from './ModalContentColorScale.svelte'
+export { default as ModalContentBarChartAxes } from './ModalContentBarChartAxes.svelte'
+export { default as ModalContentScarfPlotClip } from './ModalContentScarfPlotClip.svelte'
+
+// Data Management Modals
+export { default as ModalContentAoiModification } from './ModalContentAoiModification.svelte'
+export { default as ModalContentParticipantModification } from './ModalContentParticipantModification.svelte'
+export { default as ModalContentStimulusModification } from './ModalContentStimulusModification.svelte'
+export { default as ModalContentParticipantsGroups } from './ModalContentParticipantsGroups.svelte'
+export { default as ModalContentParticipantsGroupsChecklist } from './ModalContentParticipantsGroupsChecklist.svelte'
