+++ src/lib/modals/shared/components/SectionHeader.svelte
