+++ src/lib/modals/shared/components/index.ts
@@ -1,0 +1,7 @@
+// Main Modal Component
+export { default as Modal } from './Modal.svelte'
+
+// Shared Components
+export { default as SectionHeader } from './SectionHeader.svelte'
+export { default as SortableTableHeader } from './SortableTableHeader.svelte'
+export { default as DownloadPlotSettings } from './DownloadPlotSettings.svelte'
